backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email-validator]==2.6.1
email-validator>=2.0.0
pydantic-settings==2.1.0
httpx==0.26.0
python-multipart==0.0.9
google-auth==2.35.0
requests==2.32.5
